requirements.txt
# Core CLI and utilities
click
tqdm
python-dotenv
psutil>=5.9.0

# AI/ML libraries - pinned versions for stability
google-generativeai>=0.3.0
langchain-google-genai>=0.3.0
langchain-community>=0.3.25
faiss-cpu>=1.7.0

# PDF processing - PyPDF2 is the default (ultra-fast)
PyPDF2>=3.0.0

# Enhanced PDF processing (optional - use --enhanced or --hi-res flags)
# unstructured[pdf] provides better formatting but is much slower
unstructured[pdf]>=0.10.0

# Alternative lightweight option:
# pdfplumber>=0.9.0